fix: handle NaN y_min/y_max in DataStore and fix mksqlite test expectations
- build_store_mex: add NaN-safe fallback after SIMD min/max — rescans
  with scalar loop skipping NaN, uses Inf/-Inf sentinels for all-NaN
  chunks (SQLite converts NaN to NULL, violating NOT NULL constraints)
- FastPlotDataStore.initSqlite: same NaN→Inf/-Inf handling in MATLAB path
- TestMksqliteEdgeCases: fix int32 round-trip expectations — mksqlite
  binds scalar int32 as SQLite REAL, so it comes back as double

+++ tests/suite/TestMksqliteEdgeCases.m
@@ -228,8 +228,10 @@ classdef TestMksqliteEdgeCases < matlab.unittest.TestCase
             end
             res = mksqlite(testCase.db, 'SELECT val FROM multi WHERE id > 2 ORDER BY id');
             testCase.verifyEqual(numel(res), 3, 'multiRow: 3 results');
-            testCase.verifyEqual(res(1).val, int32(30), 'multiRow: first');
-            testCase.verifyEqual(res(3).val, int32(50), 'multiRow: last');
+            % mksqlite binds scalar numerics as SQLite REAL, so int32
+            % round-trips as double.  Compare values, not class.
+            testCase.verifyEqual(res(1).val, 30, 'multiRow: first');
+            testCase.verifyEqual(res(3).val, 50, 'multiRow: last');
         end
 
         function testZeroRowQuery(testCase)
@@ -250,7 +252,8 @@ classdef TestMksqliteEdgeCases < matlab.unittest.TestCase
             db2 = mksqlite('open', testCase.dbPath);
             mksqlite(db2, 'typedBLOBs', 2);
             res = mksqlite(db2, 'SELECT val FROM multi3 WHERE id=1');
-            testCase.verifyEqual(res.val, int32(10), 'reopen: data persists');
+            % Scalar int32 round-trips as double (see testScalarInt32).
+            testCase.verifyEqual(res.val, 10, 'reopen: data persists');
             mksqlite(db2, 'close');
             % Re-open for teardown
             testCase.db = mksqlite('open', testCase.dbPath);
